State, Summoner and My Account are extracted to separate modules

+++ data-dragon-api/src/main/java/com/tsarsprocket/reportmid/data_dragon_api/data/DataDragon.kt
@@ -21,7 +21,7 @@ interface DataDragon {
         val items: List<Item>
         fun getRunePathById(id: Int): RunePath
         fun getPerkById(id: Int): Perk
-        fun getChampionById(id: Int): Champion
+        fun getChampionById(id: Long): Champion
         fun getSummonerSpellById(id: Long): SummonerSpell
         fun getItemById(id: Int): Item
     }
